refactor(decorations): move provider code
Move decoration provider code to a separate file.
