+++ src/nvim/memory.c
@@ -10,7 +10,7 @@
 
 #include "nvim/api/extmark.h"
 #include "nvim/context.h"
-#include "nvim/decoration.h"
+#include "nvim/decoration_provider.h"
 #include "nvim/eval.h"
 #include "nvim/highlight.h"
 #include "nvim/lua/executor.h"
